+++ src/main/java/cc/cassian/immersiveoverlays/ModClient.java
@@ -10,7 +10,7 @@ public class ModClient {
     public static final String MOD_ID = "immersiveoverlays";
     public static final String MOD_NAME = "Immersive Overlays";
     public static final Logger LOGGER = LogManager.getLogger(MOD_NAME);
-    public static final int CONFIG_VERSION = 0;
+    public static final int CONFIG_VERSION = 1;
     public static String loader;
     public static void init(String loader) {
         ModConfig.load();