+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Support for seasons from Ecliptic Seasons and TerraFirmaCraft.
+
 ### Fixed
 - Missing config translation key.